+++ lib/coloraide/color/distance/delta_e_itp.py
@@ -7,13 +7,7 @@ import math
 
 
 def distance(color1, color2, scalar=720, **kwargs):
-    """
-    Delta E 1976 color distance formula.
-
-    http://www.brucelindbloom.com/index.html?Eqn_DeltaE_CIE76.html
-
-    Basically this is Euclidean distance in the Lab space.
-    """
+    """Delta E ITP color distance formula."""
 
     i1, t1, p1 = color1.convert('ictcp').coords()
     i2, t2, p2 = color2.convert('ictcp').coords()