+++ lib/coloraide/color/gamut/__init__.py
@@ -52,7 +52,7 @@ class Gamut:
 
         # If we are perfectly in gamut, don't waste time fitting, just normalize hues.
         # If out of gamut, apply mapping/clipping/etc.
-        c._space._coords = norm_angles(c) if c.in_gamut(tolerance=0.0) else func(self.clone(), c)
+        c._space._coords = norm_angles(c) if c.in_gamut(tolerance=0.0) else func(c)
 
         # Adjust "this" color
         return this.update(c)