+++ lib/coloraide/spaces/display_p3.py
@@ -1,6 +1,5 @@
 """Display-p3 color class."""
 from ..spaces import RE_DEFAULT_MATCH
-from ..spaces import _cat
 from .srgb import SRGB, lin_srgb, gam_srgb
 from .xyz import XYZ
 from .. import util
@@ -53,16 +52,16 @@ class DisplayP3(SRGB):
 
     SPACE = "display-p3"
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space=SPACE))
-    WHITE = _cat.WHITES["D65"]
+    WHITE = "D65"
 
     @classmethod
-    def _to_xyz(cls, rgb):
+    def _to_xyz(cls, parent, rgb):
         """To XYZ."""
 
-        return _cat.chromatic_adaption(cls.white(), XYZ.white(), lin_p3_to_xyz(lin_p3(rgb)))
+        return parent.chromatic_adaptation(cls.WHITE, XYZ.WHITE, lin_p3_to_xyz(lin_p3(rgb)))
 
     @classmethod
-    def _from_xyz(cls, xyz):
+    def _from_xyz(cls, parent, xyz):
         """From XYZ."""
 
-        return gam_p3(xyz_to_lin_p3(_cat.chromatic_adaption(XYZ.white(), cls.white(), xyz)))
+        return gam_p3(xyz_to_lin_p3(parent.chromatic_adaptation(XYZ.WHITE, cls.WHITE, xyz)))