+++ lib/coloraide/spaces/lab.py
@@ -1,6 +1,5 @@
 """Lab class."""
 from ..spaces import Space, RE_DEFAULT_MATCH, GamutUnbound, Percent
-from . import _cat
 from .xyz import XYZ
 from .. import util
 import re
@@ -12,7 +11,7 @@ RATIO2 = 108 / 841
 RATIO3 = 841 / 108
 
 
-def lab_to_xyz(lab):
+def lab_to_xyz(lab, white):
     """
     Convert Lab to D50-adapted XYZ.
 
@@ -34,14 +33,14 @@ def lab_to_xyz(lab):
     ]
 
     # Compute XYZ by scaling `xyz` by reference `white`
-    return util.multiply(xyz, Lab.white())
+    return util.multiply(xyz, white)
 
 
-def xyz_to_lab(xyz):
+def xyz_to_lab(xyz, white):
     """Assuming XYZ is relative to D50, convert to CIE Lab from CIE standard."""
 
     # compute `xyz`, which is XYZ scaled relative to reference white
-    xyz = util.divide(xyz, Lab.white())
+    xyz = util.divide(xyz, white)
     # Compute `fx`, `fy`, and `fz`
     fx, fy, fz = [util.cbrt(i) if i > EPSILON3 else (RATIO3 * i) + RATIO1 for i in xyz]
 
@@ -105,16 +104,16 @@ class Lab(LabBase):
 
     SPACE = "lab"
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space=SPACE))
-    WHITE = _cat.WHITES["D50"]
+    WHITE = "D50"
 
     @classmethod
-    def _to_xyz(cls, lab):
+    def _to_xyz(cls, parent, lab):
         """To XYZ."""
 
-        return _cat.chromatic_adaption(cls.white(), XYZ.white(), lab_to_xyz(lab))
+        return parent.chromatic_adaptation(cls.WHITE, XYZ.WHITE, lab_to_xyz(lab, cls.white()))
 
     @classmethod
-    def _from_xyz(cls, xyz):
+    def _from_xyz(cls, parent, xyz):
         """From XYZ."""
 
-        return xyz_to_lab(_cat.chromatic_adaption(XYZ.white(), cls.white(), xyz))
+        return xyz_to_lab(parent.chromatic_adaptation(XYZ.WHITE, cls.WHITE, xyz), cls.white())