+++ lib/coloraide/spaces/lab_d65.py
@@ -1,6 +1,5 @@
 """Lab D65 class."""
 from ..spaces import RE_DEFAULT_MATCH
-from . import _cat
 from .xyz import XYZ
 from .lab import LabBase, lab_to_xyz, xyz_to_lab
 import re
@@ -11,16 +10,16 @@ class LabD65(LabBase):
 
     SPACE = "lab-d65"
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space=SPACE))
-    WHITE = _cat.WHITES["D65"]
+    WHITE = "D65"
 
     @classmethod
-    def _to_xyz(cls, lab):
+    def _to_xyz(cls, parent, lab):
         """To XYZ."""
 
-        return _cat.chromatic_adaption(cls.white(), XYZ.white(), lab_to_xyz(lab))
+        return parent.chromatic_adaptation(cls.WHITE, XYZ.WHITE, lab_to_xyz(lab, cls.white()))
 
     @classmethod
-    def _from_xyz(cls, xyz):
+    def _from_xyz(cls, parent, xyz):
         """From XYZ."""
 
-        return xyz_to_lab(_cat.chromatic_adaption(XYZ.white(), cls.white(), xyz))
+        return xyz_to_lab(parent.chromatic_adaptation(XYZ.WHITE, cls.WHITE, xyz), cls.white())