+++ lib/coloraide/spaces/srgb_linear.py
@@ -1,6 +1,5 @@
 """SRGB Linear color class."""
 from ..spaces import RE_DEFAULT_MATCH
-from . import _cat
 from .srgb import SRGB, lin_srgb_to_xyz, xyz_to_lin_srgb, lin_srgb, gam_srgb
 from .xyz import XYZ
 import re
@@ -11,28 +10,28 @@ class SRGBLinear(SRGB):
 
     SPACE = "srgb-linear"
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space=SPACE))
-    WHITE = _cat.WHITES["D65"]
+    WHITE = "D65"
 
     @classmethod
-    def _to_srgb(cls, rgb):
+    def _to_srgb(cls, parent, rgb):
         """Linear sRGB to sRGB."""
 
         return gam_srgb(rgb)
 
     @classmethod
-    def _from_srgb(cls, rgb):
+    def _from_srgb(cls, parent, rgb):
         """sRGB to linear sRGB."""
 
         return lin_srgb(rgb)
 
     @classmethod
-    def _to_xyz(cls, rgb):
+    def _to_xyz(cls, parent, rgb):
         """SRGB Linear to XYZ."""
 
-        return _cat.chromatic_adaption(cls.white(), XYZ.white(), lin_srgb_to_xyz(rgb))
+        return parent.chromatic_adaptation(cls.WHITE, XYZ.WHITE, lin_srgb_to_xyz(rgb))
 
     @classmethod
-    def _from_xyz(cls, xyz):
+    def _from_xyz(cls, parent, xyz):
         """XYZ to SRGB Linear."""
 
-        return xyz_to_lin_srgb(_cat.chromatic_adaption(XYZ.white(), cls.white(), xyz))
+        return xyz_to_lin_srgb(parent.chromatic_adaptation(XYZ.WHITE, cls.WHITE, xyz))